+++ engine/src/main/java/net/jqwik/engine/discovery/specs/PropertyDiscoverySpec.java
@@ -1,9 +1,6 @@
 package net.jqwik.engine.discovery.specs;
 
-import java.lang.annotation.*;
 import java.lang.reflect.*;
-import java.util.*;
-import java.util.stream.*;
 
 import org.junit.platform.engine.support.hierarchical.Node.*;
 
@@ -11,6 +8,8 @@ import net.jqwik.engine.discovery.predicates.*;
 import net.jqwik.engine.support.*;
 
 import static org.junit.platform.commons.support.ModifierSupport.*;
+
+import static net.jqwik.engine.descriptor.DiscoverySupport.*;
 
 public class PropertyDiscoverySpec implements DiscoverySpec<Method> {
 	private final IsProperty isProperty = new IsProperty();
@@ -26,23 +25,14 @@ public class PropertyDiscoverySpec implements DiscoverySpec<Method> {
 			return SkipResult.skip("A @Property method must not be static");
 		if (hasJupiterAnnotation(candidate))
 			return SkipResult.skip(String.format(
-				"A @Property method must not have Jupiter annotations: %s",
-				JqwikStringSupport.displayString(findJupiterAnnotations(candidate))
+				"A @Property method must not have JUnit annotations: %s",
+				JqwikStringSupport.displayString(findJUnitAnnotations(candidate))
 			));
 		return SkipResult.doNotSkip();
 	}
 
 	private boolean hasJupiterAnnotation(Method candidate) {
-		return findJupiterAnnotations(candidate).size() > 0;
-	}
-
-	private List<Annotation> findJupiterAnnotations(Method candidate) {
-		return Arrays.stream(candidate.getDeclaredAnnotations())
-					 .filter(annotation -> annotation.annotationType()
-													 .getPackage()
-													 .getName()
-													 .startsWith("org.junit.jupiter"))
-					 .collect(Collectors.toList());
+		return findJUnitAnnotations(candidate).size() > 0;
 	}
 
 }